+++ src/Actions/MoveAndRefactorFileAction.php
@@ -10,13 +10,12 @@ class MoveAndRefactorFileAction
     ) {
     }
 
+    /**
+     * @param  array<string, string>  $refactor
+     */
     public function execute(string $sourcePath, string $targetPath, array $refactor = [], bool $force = false): void
     {
-        try {
-            $this->moveFileAction->execute($sourcePath, $targetPath, $force);
-        } catch (\Exception $e) {
-            dd($sourcePath, $targetPath);
-        }
+        $this->moveFileAction->execute($sourcePath, $targetPath, $force);
 
         if ($refactor) {
             $this->refactorFileAction->execute($targetPath, $refactor);