+++ src/Commands/MakeDataTransferObjectCommand.php
@@ -8,6 +8,7 @@ class MakeDataTransferObjectCommand extends DomainCommand
 {
     protected $signature = 'beyond:make:data {name} {--force}';
 
+    /** @var array<string> */
     protected $aliases = [
         'beyond:make:dto',
     ];