+++ src/Commands/MakePolicyCommand.php
@@ -3,20 +3,14 @@
 namespace AkrilliA\LaravelBeyond\Commands;
 
 use AkrilliA\LaravelBeyond\Commands\Abstracts\DomainCommand;
+use AkrilliA\LaravelBeyond\NameResolver;
+use Illuminate\Support\Str;
 
 class MakePolicyCommand extends DomainCommand
 {
     protected $signature = 'beyond:make:policy {name} {--model=} {--force}';
 
     protected $description = 'Make a new policy';
-
-    protected function getAdditionReplacements(): array
-    {
-        return [
-            '{{ modelName }}'     => $model = $this->option('model'),
-            '{{ modelVariable }}' => 'User' === $model ? 'object' : mb_strtolower($model),
-        ];
-    }
 
     protected function getStub(): string
     {
@@ -28,5 +22,19 @@ class MakePolicyCommand extends DomainCommand
     public function getType(): string
     {
         return 'Policy';
+    }
+
+    public function setup(NameResolver $fqn): void
+    {
+        if ($model = $this->option('model')) {
+            $command = new MakeModelCommand();
+            $fqn = $command->getNameResolver($fqn->getModule().'/'.$model);
+
+            $this->mergePlaceholders([
+                '{{ modelNamespace }}' => $fqn->getNamespace(),
+                '{{ modelClassName }}' => $fqn->getClassName(),
+                '{{ modelVariable }}' => Str::camel($fqn->getClassName())
+            ]);
+        }
     }
 }