+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `beyond:make:migration` command by [@regnerisch](https://github.com/regnerisch)
+- Add [phpstan](https://github.com/phpstan/phpstan) and fix errors by [@regnerisch](https://github.com/regnerisch)
+
+### Fixed
+- Fix module choice attempts by [@regnerisch](https://github.com/regnerisch)
+
+### Changed
+- Replace `pest` with `phpunit` by [@regnerisch](https://github.com/regnerisch)
 
 ## [7.0.0-beta.3]
 ### Changed